refactor create maintenance window page
